Moved a few shared API methods into a file called app/classes/api/helpers.rb.
